Make API Calls DRYer
