+++ api/ExperimentsApi.ts
@@ -1,9 +1,7 @@
 import { ExperimentBare } from '@/models/index'
-import { getExperimentsAuthInfo } from '@/utils/auth'
 
 import { ApiData } from './ApiData'
-import UnauthorizedError from './UnauthorizedError'
-import { resolveApiUrlRoot, PRODUCTION_API_URL_ROOT } from './utils'
+import { fetchApi } from './utils'
 
 /**
  * Finds all the available experiments.
@@ -13,24 +11,8 @@ import { resolveApiUrlRoot, PRODUCTION_API_URL_ROOT } from './utils'
  * @throws UnauthorizedError
  */
 async function findAll(): Promise<ExperimentBare[]> {
-  const apiUrlRoot = resolveApiUrlRoot()
-
-  let headers
-  if (apiUrlRoot === PRODUCTION_API_URL_ROOT) {
-    const accessToken = getExperimentsAuthInfo()?.accessToken
-    if (!accessToken) {
-      throw new UnauthorizedError()
-    }
-    headers = new Headers({ Authorization: `Bearer ${accessToken}` })
-  }
-
-  const fetchUrl = `${apiUrlRoot}/experiments`
-  return fetch(fetchUrl, {
-    method: 'GET',
-    headers: headers,
-  })
-    .then((response) => response.json())
-    .then((result) => result.experiments.map((apiData: ApiData) => new ExperimentBare(apiData)))
+  const result = await fetchApi('GET', '/experiments')
+  return result.experiments.map((apiData: ApiData) => new ExperimentBare(apiData))
 }
 
 const ExperimentsApi = {
